PH-21 # extract TimeAgo out of Event component
