+++ src/components/TimeAgo/TimeAgo.css
@@ -1,0 +1,1 @@
+.self {}
